Move slow tests
